week9/assignment: todo app - restructuring updates and performance optimizations
