+++ Week-9/todo/src/components/atoms/appHeader/AppHeader.jsx
@@ -1,0 +1,17 @@
+import PropTypes from 'prop-types';
+
+import { StyledHeader } from './appHeader.styled';
+
+const AppHeader = ({ children }) => (
+    <StyledHeader>{children}</StyledHeader>
+);
+
+AppHeader.propTypes = {
+    children: PropTypes.element,
+};
+
+AppHeader.defaultProps = {
+    children: <></>,
+};
+
+export default AppHeader;
