--- Week-9/todo/src/components/atoms/header/index.js
@@ -1,1 +1,0 @@
-export { default as Header } from "./Header";
