+++ Week-9/todo/src/components/atoms/index.js
@@ -1,7 +1,7 @@
-export * from "./brandLabel";
-export * from "./button";
-export * from "./header";
-export * from "./input";
-export * from "./inputHint";
-export * from "./label";
 export * from "./text";
+export * from "./input";
+export * from "./label";
+export * from "./button";
+export * from "./appHeader";
+export * from "./inputHint";
+export * from "./brandLabel";
