+++ Week-9/todo/src/pages/home/reducer/reducer.js
@@ -2,7 +2,6 @@ import { handleAddTodo, handleDeleteTodo, handleToggleTodo } from "./helpers";
 
 const INITIAL_STATE = {
     errorMsg: '',
-    newTodo: '',
     todoList: [],
     todoListKey: 0,
 };
@@ -14,9 +13,6 @@ const reducer = (prevState, action) => {
 
         case "DELETE_TODO":
             return handleDeleteTodo(prevState, action);
-
-        case "SET_INPUT":
-            return { ...prevState, errorMsg: '', newTodo: action.payload?.value };
 
         case "SET_TODO_LIST":
             return { ...prevState, todoList: action.payload?.todoList }
